requirements.txt
Flask==3.1.2
openai>=1.0.0
python-dotenv==1.1.1
python-docx==1.1.2
openpyxl==3.1.5
xlrd==2.0.1
